[~] functional todolist without edit

+++ website/src/components/CTodo/Ctodo.js
@@ -1,20 +1,28 @@
 import React, { Component }from 'react';
+import moment from 'moment';
 import uuid from 'uuid';
+import {prints, Donotprint, isprintable} from "../../utils"
 
 class Todo extends Component {
     constructor(props) {
         super(props);
 
-        this.input=React.createRef()        
+        this.input=React.createRef() 
+        this.description=React.createRef() 
+        this.endDate=React.createRef()       
         this.state = {
             list: [],
+            print: false
         }
     }
     addtask=() => {
         const Items={
             if:uuid.v4(),
             value: this.input.current.value,
-            Date: new Date().toUTCString()
+            Date: new Date().toUTCString(),
+            description: this.description.current.value,
+            endDate: this.endDate.current.value,
+            states: "On going"
         };
 
         if (localStorage.getItem('list')==null) {
@@ -29,20 +37,47 @@ class Todo extends Component {
         this.setState({
             list:JSON.parse(localStorage.getItem('list'))
         });
+        window.location.reload();
     }
 
 
     componentDidMount() {
         const list= window.localStorage.getItem('list');
         const parsedList = JSON.parse(list);
+        var i = 0
+        var date = moment()
+        const now = moment()
 
         if (list==null) {
             return false
-        } else {
+        } else if (isprintable()===false){
+            while (i < parsedList.length) {
+                date = moment(parsedList[i].endDate)
+                if (date.isAfter(now)) {
+                    console.log("isAfter")
+                } else {
+                    parsedList.splice(i, 1)
+                    i = i - 1
+                }
+                i = i + 1;
+            }
             this.setState({
                 list: parsedList,
             })
-            console.log(this.state.list)
+        } else {
+            while (i < parsedList.length) {
+                date = moment(parsedList[i].endDate)
+                if (date.isAfter(now)) {
+                    console.log("isAfter")
+                } else {
+                    console.log(parsedList[i].states)
+                    parsedList[i].states = "End"
+                }
+                i = i + 1;
+            }
+            this.setState({
+                list: parsedList,
+            })
         }
     }
 
@@ -53,6 +88,16 @@ class Todo extends Component {
         this.setState({list:listValue})
         localStorage.setItem('list',JSON.stringify(listValue))
     }
+
+    changePrint=()=> {
+        console.log(isprintable)
+        if (isprintable() === true) {
+            Donotprint()
+        } else {
+            prints()
+        }
+        window.location.reload();
+    }
     
     render () {
         return (
@@ -60,14 +105,24 @@ class Todo extends Component {
                 <h1>Todo app</h1>
                 <hr/>
                 <div className="container">
-                    <input type="text" placeholder="Addtask" ref={this.input}/>
+                    <input type="text" placeholder="Task Name" ref={this.input}/>
+                    <input type="text" placeholder="Task Description" ref={this.description}/>
+                    <input type="date" placeholder="Task ending" ref={this.endDate}/>
                     <button onClick={this.addtask} className="button">Add</button>
+                    <button onClick={this.changePrint}> Change Print</button>
                     <ol>
                         {
                             this.state.list.map((item,index)=>
                             {
                                 return(<li key={item.id}> {item.value}
-                                    <button className="button" type="button" value="delete" data-key={index} onClick={this.deleteItem}>
+                                    <p>{item.description}</p>
+                                    <p> This task end on {item.endDate}</p>
+                                    <p>{item.states}</p>
+                                    <p>{item.Date}</p>
+                                    <button className="button" type="button" value="delete" data-key={index}>
+                                        Edit
+                                    </button>
+                                    <button className="button" type="button" value="delete" data-key={index} onClick={(e) => {if (window.confirm('Are you sure you wish to delete this item?')) this.deleteItem(e)}}>
                                         Delete
                                     </button>
                                 </li>)
